tests/wpt/web-platform-tests/css/css-ui/parsing/outline-offset-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: parsing outline-offset with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#outline-offset">
<meta name="assert" content="outline-offset supports the full grammar '<length>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("outline-offset", "0", "0px");
test_valid_value("outline-offset", "1px");
test_valid_value("outline-offset", "2em");
test_valid_value("outline-offset", "calc(3rem + 4vw)");
</script>
</body>
</html>
